Moved functions from main to utility
